pluginmind_backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.6
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.1
fastapi==0.116.1
frozenlist==1.7.0
gunicorn==23.0.0
google-auth==2.24.0
gotrue==2.9.0
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
multidict==6.6.3
openai==1.97.1
packaging==25.0
passlib==1.7.4
postgrest==0.17.2
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pytest==8.3.4
pytest-asyncio==0.24.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.9
realtime==2.7.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlmodel==0.0.24
starlette==0.47.2
storage3==0.8.2
supabase==2.9.1
supafunc==0.6.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
xai-sdk==1.0.1
yarl==1.20.1
